+++ plugins/system/cwgears/fields/csscom.php
@@ -11,7 +11,7 @@ defined('_JEXEC') or die('Restricted access');
  * @license             GNU/GPL, see /assets/en-GB.license.txt
  * @copyright           Copyright (c) 2015 Steven Palmer All rights reserved.
  *
- * CoalaWeb Contact is free software: you can redistribute it and/or modify
+ * CoalaWeb Gears is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
  * the Free Software Foundation, either version 3 of the License, or
  * (at your option) any later version.
@@ -53,6 +53,6 @@ class JFormFieldCsscom extends CWElementCsscom {
 
 class JElementCsscom extends CWElementCsscom {
 
-    var $_name = 'com';
+    var $_name = 'csscom';
 
 }